requirements.txt
asgiref==3.3.4
astroid==2.5.6
autopep8==1.5.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.13
django-allauth==0.50.0
django-crispy-forms==1.11.2
django-environ==0.8.1
django-filter==21.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-taggit==1.4.0
djangorestframework==3.13.1
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
isort==5.8.0
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.6.0
MarkupSafe==2.1.1
mccabe==0.6.1
mysql-connector-python==8.0.29
mysqlclient==2.1.0
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==21.3
Pillow==9.1.0
protobuf==3.20.1
psycopg2-binary==2.9.3
pycodestyle==2.7.0
pycparser==2.21
PyJWT==2.4.0
pylint==2.8.2
pyparsing==3.0.9
python-sql==1.4.0
python3-openid==3.2.0
pytz==2021.1
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.0.0
wrapt==1.12.1
